Split convolve.cc
In some builds it takes enormous amount of time to compile.
Let it compile at least in parallel.
